[hotfix][connector-files][connector-hive] Move Hive specific configs out of FileSystemConnectorOptions
This closes #18177.
