+++ flink-connectors/flink-connector-hive/src/test/java/org/apache/flink/connectors/hive/HiveDynamicTableFactoryTest.java
@@ -19,7 +19,7 @@
 package org.apache.flink.connectors.hive;
 
 import org.apache.flink.configuration.Configuration;
-import org.apache.flink.connector.file.table.FileSystemConnectorOptions.PartitionOrder;
+import org.apache.flink.connectors.hive.HiveOptions.PartitionOrder;
 import org.apache.flink.table.api.EnvironmentSettings;
 import org.apache.flink.table.api.SqlDialect;
 import org.apache.flink.table.api.TableEnvironment;
@@ -44,14 +44,14 @@ import org.junit.Test;
 
 import java.time.Duration;
 
-import static org.apache.flink.connector.file.table.FileSystemConnectorOptions.LOOKUP_JOIN_CACHE_TTL;
 import static org.apache.flink.connector.file.table.FileSystemConnectorOptions.PARTITION_TIME_EXTRACTOR_CLASS;
 import static org.apache.flink.connector.file.table.FileSystemConnectorOptions.PARTITION_TIME_EXTRACTOR_KIND;
-import static org.apache.flink.connector.file.table.FileSystemConnectorOptions.STREAMING_SOURCE_CONSUME_START_OFFSET;
-import static org.apache.flink.connector.file.table.FileSystemConnectorOptions.STREAMING_SOURCE_ENABLE;
-import static org.apache.flink.connector.file.table.FileSystemConnectorOptions.STREAMING_SOURCE_MONITOR_INTERVAL;
-import static org.apache.flink.connector.file.table.FileSystemConnectorOptions.STREAMING_SOURCE_PARTITION_INCLUDE;
-import static org.apache.flink.connector.file.table.FileSystemConnectorOptions.STREAMING_SOURCE_PARTITION_ORDER;
+import static org.apache.flink.connectors.hive.HiveOptions.LOOKUP_JOIN_CACHE_TTL;
+import static org.apache.flink.connectors.hive.HiveOptions.STREAMING_SOURCE_CONSUME_START_OFFSET;
+import static org.apache.flink.connectors.hive.HiveOptions.STREAMING_SOURCE_ENABLE;
+import static org.apache.flink.connectors.hive.HiveOptions.STREAMING_SOURCE_MONITOR_INTERVAL;
+import static org.apache.flink.connectors.hive.HiveOptions.STREAMING_SOURCE_PARTITION_INCLUDE;
+import static org.apache.flink.connectors.hive.HiveOptions.STREAMING_SOURCE_PARTITION_ORDER;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
@@ -87,7 +87,8 @@ public class HiveDynamicTableFactoryTest {
         Configuration configuration = new Configuration();
         tableSource.catalogTable.getOptions().forEach(configuration::setString);
         assertEquals(
-                PartitionOrder.PARTITION_NAME, configuration.get(STREAMING_SOURCE_PARTITION_ORDER));
+                HiveOptions.PartitionOrder.PARTITION_NAME,
+                configuration.get(STREAMING_SOURCE_PARTITION_ORDER));
 
         // test table can't be selected when set 'streaming-source.partition.include' to 'latest'
         tableEnv.executeSql(
@@ -122,7 +123,7 @@ public class HiveDynamicTableFactoryTest {
         Configuration configuration1 = new Configuration();
         hiveTableSource3.catalogTable.getOptions().forEach(configuration1::setString);
         PartitionOrder partitionOrder1 = configuration1.get(STREAMING_SOURCE_PARTITION_ORDER);
-        assertEquals(PartitionOrder.PARTITION_NAME, partitionOrder1);
+        assertEquals(HiveOptions.PartitionOrder.PARTITION_NAME, partitionOrder1);
 
         // test deprecated option key 'streaming-source.consume-order' and new key
         // 'streaming-source.partition-order'
@@ -139,7 +140,7 @@ public class HiveDynamicTableFactoryTest {
         Configuration configuration2 = new Configuration();
         hiveTableSource.catalogTable.getOptions().forEach(configuration2::setString);
         PartitionOrder partitionOrder2 = configuration2.get(STREAMING_SOURCE_PARTITION_ORDER);
-        assertEquals(PartitionOrder.PARTITION_TIME, partitionOrder2);
+        assertEquals(HiveOptions.PartitionOrder.PARTITION_TIME, partitionOrder2);
     }
 
     @Test
@@ -170,7 +171,8 @@ public class HiveDynamicTableFactoryTest {
         Configuration configuration = new Configuration();
         lookupTableSource.catalogTable.getOptions().forEach(configuration::setString);
         assertEquals(
-                configuration.get(STREAMING_SOURCE_PARTITION_ORDER), PartitionOrder.PARTITION_NAME);
+                configuration.get(STREAMING_SOURCE_PARTITION_ORDER),
+                HiveOptions.PartitionOrder.PARTITION_NAME);
 
         // test lookup with partition-time extractor options
         tableEnv.executeSql(
@@ -198,7 +200,7 @@ public class HiveDynamicTableFactoryTest {
 
         assertEquals(
                 configuration1.get(STREAMING_SOURCE_PARTITION_ORDER),
-                PartitionOrder.PARTITION_TIME);
+                HiveOptions.PartitionOrder.PARTITION_TIME);
         assertEquals(configuration1.get(PARTITION_TIME_EXTRACTOR_KIND), "custom");
         assertEquals(configuration1.get(PARTITION_TIME_EXTRACTOR_CLASS), "path.to..TimeExtractor");
