+++ flink-connectors/flink-connector-hive/src/test/java/org/apache/flink/connectors/hive/HiveLookupJoinITCase.java
@@ -19,7 +19,6 @@
 package org.apache.flink.connectors.hive;
 
 import org.apache.flink.api.common.typeutils.TypeSerializer;
-import org.apache.flink.connector.file.table.FileSystemConnectorOptions;
 import org.apache.flink.connector.file.table.PartitionFetcher;
 import org.apache.flink.connector.file.table.PartitionReader;
 import org.apache.flink.table.api.EnvironmentSettings;
@@ -52,10 +51,10 @@ import java.util.List;
 
 import static org.apache.flink.connector.file.table.FileSystemConnectorOptions.PARTITION_TIME_EXTRACTOR_KIND;
 import static org.apache.flink.connector.file.table.FileSystemConnectorOptions.PARTITION_TIME_EXTRACTOR_TIMESTAMP_PATTERN;
-import static org.apache.flink.connector.file.table.FileSystemConnectorOptions.STREAMING_SOURCE_ENABLE;
-import static org.apache.flink.connector.file.table.FileSystemConnectorOptions.STREAMING_SOURCE_MONITOR_INTERVAL;
-import static org.apache.flink.connector.file.table.FileSystemConnectorOptions.STREAMING_SOURCE_PARTITION_INCLUDE;
-import static org.apache.flink.connector.file.table.FileSystemConnectorOptions.STREAMING_SOURCE_PARTITION_ORDER;
+import static org.apache.flink.connectors.hive.HiveOptions.STREAMING_SOURCE_ENABLE;
+import static org.apache.flink.connectors.hive.HiveOptions.STREAMING_SOURCE_MONITOR_INTERVAL;
+import static org.apache.flink.connectors.hive.HiveOptions.STREAMING_SOURCE_PARTITION_INCLUDE;
+import static org.apache.flink.connectors.hive.HiveOptions.STREAMING_SOURCE_PARTITION_ORDER;
 import static org.junit.Assert.assertEquals;
 
 /** Test lookup join of hive tables. */
@@ -88,7 +87,7 @@ public class HiveLookupJoinITCase {
         tableEnv.executeSql(
                 String.format(
                         "create table bounded_table (x int, y string, z int) tblproperties ('%s'='5min')",
-                        FileSystemConnectorOptions.LOOKUP_JOIN_CACHE_TTL.key()));
+                        HiveOptions.LOOKUP_JOIN_CACHE_TTL.key()));
 
         // create the hive partitioned table
         tableEnv.executeSql(
@@ -96,7 +95,7 @@ public class HiveLookupJoinITCase {
                         "create table bounded_partition_table (x int, y string, z int) partitioned by ("
                                 + " pt_year int, pt_mon string, pt_day string)"
                                 + " tblproperties ('%s'='5min')",
-                        FileSystemConnectorOptions.LOOKUP_JOIN_CACHE_TTL.key()));
+                        HiveOptions.LOOKUP_JOIN_CACHE_TTL.key()));
 
         // create the hive partitioned table
         tableEnv.executeSql(
@@ -115,9 +114,9 @@ public class HiveLookupJoinITCase {
                         "create table partition_table_1 (x int, y string, z int) partitioned by ("
                                 + " pt_year int, pt_mon string, pt_day string)"
                                 + " tblproperties ('%s' = 'true', '%s' = 'latest', '%s'='120min')",
-                        FileSystemConnectorOptions.STREAMING_SOURCE_ENABLE.key(),
-                        FileSystemConnectorOptions.STREAMING_SOURCE_PARTITION_INCLUDE.key(),
-                        FileSystemConnectorOptions.STREAMING_SOURCE_MONITOR_INTERVAL.key()));
+                        HiveOptions.STREAMING_SOURCE_ENABLE.key(),
+                        HiveOptions.STREAMING_SOURCE_PARTITION_INCLUDE.key(),
+                        HiveOptions.STREAMING_SOURCE_MONITOR_INTERVAL.key()));
 
         // create the hive partitioned table3 which uses 'partition-time'.
         tableEnv.executeSql(
